Moved scanning code to the framework.
